iRoad with functonal Accient reporter
Fixed the accdent reporter that sends all accident information to the
server.

Known issues are some few fields arent yet sent.
There are issues with uploading large attachments, videos and images to
the server on a slow network connection.
A selected image fle is not yet uploaded to the server.
The offence history and offence payments are not yet optimised.